console status messages

+++ index.js
@@ -3,6 +3,7 @@ const fs = require('fs');
 const directory = path.join(process.cwd(), 'out');
 
 processFiles(directory);
+console.log('The next project output was successfully fixed');
 
 function processFiles(pathName) {
   const names = fs.readdirSync(pathName);
@@ -10,7 +11,10 @@ function processFiles(pathName) {
     if (file.match(/\./)) {
       const originalContent = fs.readFileSync(path.join(pathName, file), 'utf-8');
       const replacedContent = originalContent.replace(/\/_next/g, './_next');
-      fs.writeFileSync(path.join(pathName, file), replacedContent);
+      if(originalContent !== replacedContent) {
+        fs.writeFileSync(path.join(pathName, file), replacedContent);
+        console.log(path.join(pathName, file) + ' done! :)')
+      }
     } else {
       processFiles(path.join(pathName, file));
     }
